Move Container and LocalLinker to link module
The comment said to move Container to compile, but it is directly and
exclusively used by the Linkers.  Some type hints were added and Container now
requires kwarg specification for the optional arguments.
