+++ theano/compile/sharedvalue.py
@@ -9,9 +9,9 @@ import logging
 import numpy as np
 
 from theano.gof.graph import Variable
-from theano.gof.link import Container
 from theano.gof.type import generic
 from theano.gof.utils import add_tag_trace
+from theano.link import Container
 
 
 _logger = logging.getLogger("theano.compile.sharedvalue")
