+++ theano/gof/__init__.py
@@ -5,14 +5,7 @@ from theano.gof.cc import CLinker, DualLinker, HideC, OpWiseCLinker
 from theano.gof.destroyhandler import DestroyHandler
 from theano.gof.fg import FunctionGraph, InconsistencyError, MissingInputError
 from theano.gof.graph import Apply, Constant, Variable, view_roots
-from theano.gof.link import (
-    Container,
-    Linker,
-    LocalLinker,
-    PerformLinker,
-    WrapLinker,
-    WrapLinkerMany,
-)
+from theano.gof.link import PerformLinker, WrapLinker, WrapLinkerMany
 from theano.gof.op import (
     COp,
     Op,
@@ -54,6 +47,7 @@ from theano.gof.toolbox import (
 )
 from theano.gof.type import CEnumType, EnumList, EnumType, Generic, Type, generic
 from theano.gof.utils import MethodNotDefined, hashtype, object2
+from theano.link import Container, Linker, LocalLinker
 
 
 if theano.config.cmodule__preload_cache:
