+++ theano/gof/cc.py
@@ -1214,7 +1214,7 @@ class CLinker(link.Linker):
                 for input, storage in zip(self.fgraph.inputs, input_storage)
             ],
             [
-                link.Container(output, storage, True)
+                link.Container(output, storage, readonly=True)
                 for output, storage in zip(self.fgraph.outputs, output_storage)
             ],
             error_storage,
@@ -1987,7 +1987,7 @@ class OpWiseCLinker(link.LocalLinker):
                 for input, storage in zip(fgraph.inputs, input_storage)
             ],
             [
-                link.Container(output, storage, True)
+                link.Container(output, storage, readonly=True)
                 for output, storage in zip(fgraph.outputs, output_storage)
             ],
             thunks,
