+++ theano/link/__init__.py
@@ -1,0 +1,1 @@
+﻿from theano.link.basic import Container, Linker, LocalLinker
