+++ theano/link/jax/jax_linker.py
@@ -1,16 +1,16 @@
 from collections.abc import Sequence
 from warnings import warn
 
+from theano.gof import utils
 from theano.gof.graph import Constant
 from theano.gof.link import (
-    Container,
     PerformLinker,
     add_clear_storage,
     gc_helper,
     map_storage,
     streamline,
-    utils,
 )
+from theano.link import Container
 
 
 class JAXLinker(PerformLinker):
@@ -194,7 +194,7 @@ class JAXLinker(PerformLinker):
                 for input, storage in zip(fgraph.inputs, input_storage)
             ],
             [
-                Container(output, storage, True)
+                Container(output, storage, readonly=True)
                 for output, storage in zip(fgraph.outputs, output_storage)
             ],
             thunks,
